feat(auth): add bypass for unary and stream interceptors
Add a bypass mechanism to the unary and streaming interceptors in the auth
package. This allows requests to proceed directly to the handler without
going through the authentication process, which can be useful for health checks
or other scenarios where authentication is not required.feat(commonfield): introduce UnixTimestamp struct

Introduce a new UnixTimestamp struct within the commonfield package to handletimestamp management. This struct incorporates `CreatedAt` and `UpdatedAt` fields,
which are automatically set by GORM's `autoCreateTime` and `autoUpdateTime`features, respectively. The addition of this struct helps streamline timestamp
handling within models utilizing the commonfield.